+++ monidlistener/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\Cheeseball\\AppData\\Local\\Programs\\Python\\Python38-32\\python.exe"
+}
