+++ monidlistener/__init__.py
@@ -1,0 +1,7 @@
+from .monidlistener import MonIdListener
+
+__red_end_user_data_statement__ = "No personal data is stored."
+
+
+def setup(bot):
+    bot.add_cog(MonIdListener(bot))
